reorg files
move old stuff to insight file
outpoint tracking still doesn't work, but in most cases
wallet works OK without it.
Height still doesn't get set.  getting there.
